fix(frontend): merge Nlp training form and nlp sample form
