+++ frontend/src/components/nlp/components/NlpSample.tsx
@@ -184,7 +184,6 @@ export default function NlpSample() {
             { defaultValues: data },
             {
               maxWidth: "md",
-              hasButtons: false,
             },
           );
         },
